mhhh let's see what this does lulw LET'S HOPE IT WORKS HAHAHAHA

+++ src/main/java/lavalink/client/player/LavalinkPlayer.java
@@ -176,7 +176,7 @@ public class LavalinkPlayer implements IPlayer {
     @Deprecated
     public void setVolume(int volume) {
         if (filters == null && volume == 100) return;
-        getFilters().setVolume(volume / 100).commit();
+        getFilters().setVolume((float)volume / 100).commit();
     }
 
     /**
